requirements.txt
Flask~=1.1.2
Flask-SQLAlchemy~=2.4.4
Flask-Migrate~=2.5.3
psycopg2-binary~=2.8.6
arrow~=0.17.0
flask-swagger-ui~=3.36.0
pyyaml~=5.3.1
